Python/requirements.txt
flask==3.1.3
mysql-connector-python==9.1.0
